Die, Dice, and WordBuilder classes created. BoggleBoard can now create a random board.
Each Die object stores the 6 possible sides of a die and the current side. The Dice class stores a list of Die objects and instantiates them so that they are the correct dice in a boggle game, either the classic dice or the modern dice. BoggleBoard has a Dice object, and it has a list which stores which Die is at which grid location. The Die objects are numbered based on their location in the Dice object's Die list.

WordBuilder works as a stack where grid space numbers can be pushed and popped. Numbers can only be pushed if that number has not already been used and it is adjacent to the current location, which is stored. These numbers can be popped off, which will be used when searching the board for words with the prefix tree.